+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/mechanisms/Arm.java
@@ -7,6 +7,7 @@ import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.Gamepad;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.Servo;
+import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.MecanumDrive;
@@ -14,24 +15,26 @@ import org.firstinspires.ftc.teamcode.wagner.PartsMap;
 
 @Config
 public class Arm implements Mechanism {
+    public ElapsedTime _timer;
     public DcMotor left;
     public DcMotor right;
-    public Servo elbow;
-    public Servo shoulder;
+    public CRServo elbow;
+    public CRServo shoulder;
 
-    public static double elbowPos = 0;
     public static double shoulderPos = 0;
 
     @Override
     public void init(HardwareMap hardwareMap) {
+        _timer = new ElapsedTime();
         left = hardwareMap.get(DcMotor.class, PartsMap.ARM_LEFT.toString());
         right = hardwareMap.get(DcMotor.class, PartsMap.ARM_RIGHT.toString());
 
-        elbow = hardwareMap.get(Servo.class, PartsMap.ARM_ELBOW.toString());
-        shoulder = hardwareMap.get(Servo.class, PartsMap.ARM_SHOULDER.toString());
+        elbow = hardwareMap.get(CRServo.class, PartsMap.ARM_ELBOW.toString());
+        shoulder = hardwareMap.get(CRServo.class, PartsMap.ARM_SHOULDER.toString());
 
-        //TODO: REVERSE
-        right.setDirection(DcMotorSimple.Direction.REVERSE);
+        left.setDirection(DcMotorSimple.Direction.REVERSE);
+        right.setDirection(DcMotorSimple.Direction.FORWARD);
+
     }
 
     public void slidePowerThreshold(double pwr, double threshold) {
@@ -43,7 +46,18 @@ public class Arm implements Mechanism {
         right.setPower(pwr);
     }
 
-    public void setShoulder(double pos) {this.shoulder.setPosition(pos);Arm.shoulderPos=pos;}
-    public void setElbow(double pos) {this.elbow.setPosition(pos);Arm.elbowPos=pos;}
+    public void setShoulder(double pos) {
+        this.shoulder.setPower(pos);
+        Arm.shoulderPos=pos;
+    }
+    public void rawElbowPower(double power) {
+        this.elbow.setPower(power);
+    }
+
+    public void setElbowPowerFor(double power, double ms) {
+        _timer.reset();
+        while(_timer.milliseconds() < ms) rawElbowPower(power);
+        rawElbowPower(0);
+    }
 
 }