+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/mechanisms/Mechanism.java
@@ -1,11 +1,13 @@
 package org.firstinspires.ftc.teamcode.wagner.mechanisms;
 
+import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.robotcore.hardware.Gamepad;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.MecanumDrive;
 
+@Config
 public interface Mechanism {
     void init(HardwareMap hardwareMap);
 }